implemented Copy command builder

+++ core/Tasks/File.copy.fs
@@ -65,12 +65,14 @@ module CopyImpl =
 
             if not args.dryrun then
                 ensureDirCreated tofile
-                File.Copy(fullname, tofile, true)
+
+                if args.overwrite || not (File.Exists tofile) then
+                    File.Copy(fullname, tofile, true)
+                else
+                    ctx.Logger.Log Level.Info "[copy] skipped ('%s' already exists while overwrite option is off)" tofile
 
         let projectRoot = ctx.Options.ProjectRoot
         let targetDir = args.todir |> function | null -> projectRoot | s -> System.IO.Path.Combine(projectRoot, s)
-
-        // TODO overwrite
 
         let fileset =
             args
@@ -122,3 +124,23 @@ module CopyImpl =
         let! tgt = getTargetFullName()
         do! copyFile src tgt
     }
+
+    type CopyBuilder() =
+
+        [<CustomOperation("file")>]    member this.File(a :CopyArgs, value) =   {a with file = value }
+        [<CustomOperation("dir")>]     member this.Dir(a :CopyArgs, value) =    {a with dir = value}
+        [<CustomOperation("files")>]   member this.Fileset(a :CopyArgs, value)= {a with files = value}
+        [<CustomOperation("todir")>]   member this.ToDir(a :CopyArgs, value)=   {a with todir = value}
+        [<CustomOperation("verbose")>] member this.Verbose(a :CopyArgs) =        {a with verbose = true}
+        [<CustomOperation("overwrite")>] member this.Overwrite(a :CopyArgs) =    {a with overwrite = true}
+        [<CustomOperation("flatten")>] member this.Flatten(a :CopyArgs) =        {a with flatten = true}
+        [<CustomOperation("dryrun")>] member this.Dryrun(a :CopyArgs) =          {a with dryrun = true}
+
+        member this.Bind(x, f) = f x
+        member this.Yield(()) = CopyArgs.Default
+        member x.For(sq, b) = for e in sq do b e
+
+        member this.Zero() = CopyArgs.Default
+        member this.Run(args:CopyArgs) = Copy args
+
+    let copy = CopyBuilder()
